Adds expected Kernel vectorization + testing stuff
Now supporting generating input matrices of
arbitrary size. Values are generated using a
PRNG. Also, the input string for generating the
input AST is derived from that.

Fixes some bugs in the specification of the
Kernel program in our input language.

Fixes secret tainting test: Node IDs are now
dynamically obtained instead of relying on same ID
 (depends on test execution order -> fails).

Removes unnecessary FHE operators.

Added code for the vectorized variant of
the kernel program. This basically is what we did
for Patrick's thesis but not at runtime anymore,
and also includes masking to only use the batched
results that we need.

Signed-off-by: Patrick Jattke <pjattke@ethz.ch>